+++ CHANGES.md
@@ -5,6 +5,7 @@ v1.1.1 - YYYY-MM-DD
 -------------------
 
 - Fixed some clang-reported issues.
+- Fixed builds on Windows.
 
 
 v1.1.0 - 2025-12-31